# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Account:
    properties:
      created_at:
        $ref: '#/definitions/models.CustomTime'
      creds_file:
        type: string
      description:
        type: string
      id:
        type: string
      is_system_account:
        description: 新增字段
        type: boolean
      jwt_claims:
        $ref: '#/definitions/models.JSONMap'
      jwt_text:
        description: 存储实际的JWT密文
        type: string
      limits:
        $ref: '#/definitions/models.AccountLimits'
      n_key:
        type: string
      name:
        type: string
      origin_cluster_id:
        description: 创建来源集群ID
        type: string
      public_key:
        type: string
      status:
        $ref: '#/definitions/models.AccountStatus'
      updated_at:
        $ref: '#/definitions/models.CustomTime'
    required:
    - name
    type: object
...
  models.AccountListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Account'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  models.CreateAccountRequest:
    properties:
      description:
        type: string
      limits:
        $ref: '#/definitions/models.AccountLimits'
      name:
        type: string
      origin_cluster_id:
        type: string
    required:
    - name
    type: object
...
  models.JWTResponse:
    properties:
      claims: {}
      raw_token:
        type: string
    type: object
...
  models.UpdateAccountRequest:
    properties:
      description:
        type: string
      limits:
        $ref: '#/definitions/models.AccountLimits'
      name:
        type: string
      origin_cluster_id:
        type: string
      status:
        $ref: '#/definitions/models.AccountStatus'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing NATS control plane with multi-cluster support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NATS Control API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Get a list of NATS accounts with pagination and filters
      parameters:
      - description: Number of accounts to return (default 20)
        in: query
        name: limit
        type: integer
      - description: Number of accounts to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Search in account name and description
        in: query
        name: search
        type: string
      - description: Filter by account status (active/disabled)
        in: query
        name: status
        type: string
      - description: Filter by account type (system/normal)
        in: query
        name: account_type
        type: string
      - description: Filter by origin cluster ID
        in: query
        name: cluster_id
        type: string
      - description: Sort field (name/created_at/updated_at/status)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of accounts with pagination
          schema:
            $ref: '#/definitions/models.AccountListResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new NATS account with the specified configuration
      parameters:
      - description: Account creation request
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created account
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new NATS account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      description: Delete a NATS account and remove it from the system
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Cannot delete account with associated users
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete account
      tags:
      - accounts
    get:
      description: Retrieve a NATS account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account details
          schema:
            $ref: '#/definitions/models.Account'
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an existing NATS account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update request
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated account
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update account
      tags:
      - accounts
  /accounts/{id}/associate:
    post:
      consumes:
      - application/json
      description: Create bidirectional communication between two accounts
      parameters:
      - description: Source Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Association configuration
        in: body
        name: association
        required: true
        schema:
          $ref: '#/definitions/models.AccountAssociationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Association created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create cross-account association
      tags:
      - accounts
  /accounts/{id}/disable:
    post:
      description: Disable a NATS account (removes JWT from NATS server)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account disabled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Disable account
      tags:
      - accounts
  /accounts/{id}/enable:
    post:
      description: Enable a NATS account (pushes JWT to NATS server)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account enabled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Enable account
      tags:
      - accounts
  /accounts/{id}/exports:
    post:
      consumes:
      - application/json
      description: Add a subject export to an account for cross-account communication
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Export configuration
        in: body
        name: export
        required: true
        schema:
          $ref: '#/definitions/models.AddExportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated account
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add export to account
      tags:
      - accounts
  /accounts/{id}/exports/{name}:
    delete:
      description: Remove a subject export from an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Export name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated account
          schema:
            $ref: '#/definitions/models.Account'
        "404":
          description: Account or export not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove export from account
      tags:
      - accounts
  /accounts/{id}/imports:
    post:
      consumes:
      - application/json
      description: Add a subject import to an account for cross-account communication
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Import configuration
        in: body
        name: import
        required: true
        schema:
          $ref: '#/definitions/models.AddImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated account
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add import to account
      tags:
      - accounts
  /accounts/{id}/imports/{name}:
    delete:
      description: Remove a subject import from an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Import name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated account
          schema:
            $ref: '#/definitions/models.Account'
        "404":
          description: Account or import not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove import from account
      tags:
      - accounts
  /accounts/{id}/jwt:
    get:
      description: Get the JWT token and decoded claims for an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account JWT and claims
          schema:
            $ref: '#/definitions/models.JWTResponse'
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get account JWT
      tags:
      - accounts
  /accounts/{id}/sync:
    post:
      consumes:
      - application/json
      description: 手动将账户的JWT同步到选定的集群，通过异步任务处理
      parameters:
      - description: 账户ID
        in: path
        name: id
        required: true
        type: string
      - description: 同步请求
        in: body
        name: sync
        required: true
        schema:
          $ref: '#/definitions/models.ManualSyncJWTRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 同步任务创建成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 账户未找到
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 手动同步账户JWT到指定集群
      tags:
      - accounts
...
schemes:
- http
- https
swagger: "2.0"